-Added ShopData file -Changes to StatisticsShopCategory

+++ EntityDeathSystem.cpp
@@ -72,10 +72,10 @@ int EntityDeathSystem::getGoldValue(Entity& entity) const
 {
 	const auto& combatStats = entity.getComponent<CombatStatsComponent>();
 	constexpr int baseGold = 2;
-	int entityPower = ( (combatStats.cMaxHealth / 3)
+	int entityPower = static_cast<int>(( (combatStats.cMaxHealth / 3)
 		+ combatStats.cDefence * 2
 		+ combatStats.cMagicDefence * 2)
-		* (1 + combatStats.cAttackDamage / 20.f);
+		* (1 + combatStats.cAttackDamage / 20.f));
 
 	int goldValue = baseGold * Config::difficulityLevel * entityPower;
 	int minGold = static_cast<int>(goldValue * 0.9f);
